files consolidation
