+++ go2x3/api.go2x3.go
@@ -1,6 +1,7 @@
 package go2x3
 
 import (
+	"errors"
 	"io"
 
 	"github.com/art-media-platform/amp-sdk-go/stdlib/generics"
@@ -57,7 +58,6 @@ type System interface {
 // 	0,
 // 	1, 2, 3, 4, 5, 6, 7, 8,
 // |
-
 
 // func (gb *graphBuilder) EmitNextEdgeSweep(uptoShapeID uint64, tag.OctalDigit) {
 
@@ -255,3 +255,21 @@ type FactorRun struct {
 	ID    TracesID
 	Count uint32
 }
+
+// Errors
+var (
+	ErrUnmarshal          = errors.New("unmarshal failed")
+	ErrBadCatalogParam    = errors.New("bad catalog param")
+	ErrInsufficientTraces = errors.New("insufficient traces")
+	ErrBadEncoding        = errors.New("bad graph encoding")
+	ErrBadVtxID           = errors.New("bad graph vertex ID")
+	ErrMissingVtxID       = errors.New("missing vertex ID")
+	ErrBadEdge            = errors.New("bad graph edge")
+	ErrBadEdgeType        = errors.New("bad graph edge type")
+	ErrBrokenEdges        = errors.New("bad or inconsistent graph edge configuration")
+	ErrViolates2x3        = errors.New("graph is not a valid 2x3")
+	ErrVtxExpected        = errors.New("vertex ID expected")
+	ErrSitesExceeded      = errors.New("number of loops and edges exceeds 3")
+	ErrNilGraph           = errors.New("nil graph")
+	ErrInvalidVtxID       = errors.New("invalid vertex or group ID")
+)
