--- go2x3/api.go2x3.proto
@@ -1,6 +1,0 @@
-syntax = "proto3";
-package go2x3;
-
-
-
-
