--- go2x3/api.graph.go
@@ -1,2 +1,0 @@
-package go2x3
-
