+++ lib2x3/graph-legacy/edge.go
