+++ lib2x3/graph-legacy/vertex.go
