+++ lib2x3/graph-walker/api.walker.go
