+++ lib2x3/graph-walker/walker.go
